+++ src/shared/replication/rules.rs
@@ -7,7 +7,7 @@ use bevy::{ecs::archetype::Archetype, prelude::*};
 use serde::{Serialize, de::DeserializeOwned};
 
 use super::registry::{ReplicationRegistry, receive_fns::MutWrite};
-use crate::prelude::*;
+use crate::{prelude::*, shared::replication::diff::Diffable};
 use component::{BundleRules, ComponentRule, IntoComponentRules};
 use filter::{FilterRule, FilterRules};
 
@@ -33,6 +33,28 @@ pub trait AppRuleExt {
         C: Component<Mutability: MutWrite<C>> + Serialize + DeserializeOwned,
     {
         self.replicate_once_filtered::<C, ()>()
+    }
+
+    /// Like [`Self::replicate`], but instead of re-sending the entire component
+    /// when it changes, this sends a list of patches.
+    ///
+    /// All mutations should be performed through [`DiffEntityExt::apply_patch`],
+    /// which records the changes.
+    ///
+    /// See [`Diffable`] for more details.
+    fn replicate_diff<C>(&mut self) -> &mut Self
+    where
+        C: Diffable,
+    {
+        self.replicate_diff_filtered::<C, ()>()
+    }
+
+    /// Like [`Self::replicate_diff`], but also adds filters like [`Self::replicate_filtered`].
+    fn replicate_diff_filtered<C, F: FilterRules>(&mut self) -> &mut Self
+    where
+        C: Diffable,
+    {
+        self.replicate_with_filtered::<_, F>(RuleFns::<C>::new_diff())
     }
 
     /// Like [`Self::replicate`], but converts the component into `T` before serialization